NUNCA USAR FLOAT PARA VALORES MONETÁRIOS - Alterando para int64 formatos para dinheiro

+++ internal/dto/vendas.go
@@ -3,13 +3,13 @@ package dto
 type ItemVendaRequest struct {
 	ProdutoID     uint    `json:"produto_id"`
 	Quantidade    int     `json:"quantidade"`
-	PrecoUnitario float64 `json:"preco_unitario"`
-	Desconto      float64 `json:"desconto"`
+	PrecoUnitario int64   `json:"preco_unitario"`
+	Desconto      int64   `json:"desconto"`
 }
 
 type CriarVendaRequest struct {
 	ClienteID     *uint              `json:"cliente_id"`     // nil = venda sem identificação
-	DescontoTotal float64            `json:"desconto_total"` // opcional
+	DescontoTotal int64   `json:"desconto_total"` // opcional
 	Itens         []ItemVendaRequest `json:"itens"`
 }
